+++ inkcyauto/util.py
@@ -1,6 +1,7 @@
 """
 Misc. utility functions.
 """
+import os
 
 def get_filepath(base, path):
     """
